fix(build): clear all MSVC /W4 warnings (C4244, C4090, C4206)
C4244: cast ptrdiff_t->int in dia_find_index. C4090: project_find_task/project_find_milestone now take const Project* (idiomatic search-fn pattern, cf. strchr) so const callers don't drop qualifiers. C4206: removed dead scheduler.c/.h stubs (superseded by algorithms, not included anywhere) from vcxproj.

Rebuild: 0 warnings, 0 errors.

+++ fin-project/src/dynamic_int_array.c
@@ -42,7 +42,7 @@ int  dia_find_index(const DynamicIntArray* a, int value)
 {
     int *fnd_ptr = bsearch(&value, a->data, (size_t)a->count, sizeof(value), cmp);
     if (fnd_ptr == NULL) return -1;
-    return fnd_ptr - a->data;
+    return (int)(fnd_ptr - a->data);
 }
 int dia_remove(DynamicIntArray *a, int value) {
     int i;
